UPD API version to v2
Only difference between v1 and v2 of the API is the output of error
messages. These are now presented as strings instead of integers.

Along with the version bump the version path prefix was removed from the
URL read from the configuration file and the default API URL. This
concept turned out to be very flawed as we now have to tell our existing
clients to edit their configuration files one they update their clients.